refactored requests isOwner permissioning into a higher order component
